+++ src/utilities/export/export.ts
@@ -10,26 +10,26 @@ export function getUnixTimestamp() {
 }
 
 export function exportTsvFile(imageName: string, plotName: string, content: string) {
-    const tsvData = `data:text/tab-separated-values;charset=utf-8,${content}\n`;
+    const tsvData = `data:text/tab-separated-values;charset=utf-8,${content}\n`.trim();
     const dataURL = encodeURI(tsvData).replace(/#/g, "%23");
 
     const a = document.createElement("a") as HTMLAnchorElement;
     a.href = dataURL;
 
-    a.download = `${imageName}-${plotName.replace(" ", "-")}-${getTimestamp()}.tsv`;
+    a.download = `${imageName.replaceAll(" ", "__")}-${plotName.replaceAll(" ", "-")}-${getTimestamp()}.tsv`;
     a.dispatchEvent(new MouseEvent("click"));
 
     return null;
 }
 
 export function exportTxtFile(fileName: string, content: string) {
-    const txtData = `data:text/plain;charset=utf-8,${content}\n`;
+    const txtData = `data:text/plain;charset=utf-8,${content}\n`.trim();
     const dataURL = encodeURI(txtData).replace(/#/g, "%23");
 
     const a = document.createElement("a") as HTMLAnchorElement;
     a.href = dataURL;
 
-    a.download = `${fileName}.txt`;
+    a.download = `${fileName.replaceAll(" ", "__")}.txt`;
     a.dispatchEvent(new MouseEvent("click"));
 
     return null;